Refactor meta and app server

+++ src/main/java/de/umr/raft/raftlogreplicationdemo/models/CreatePartitionRequest.java
@@ -3,6 +3,6 @@ package de.umr.raft.raftlogreplicationdemo.models;
 // TODO obsolete, use RegisterPartitionRequest
 public interface CreatePartitionRequest {
 
-    abstract long getPartitionsCount();
+    abstract int getPartitionsCount();
 
 }
